refactor(theme-default): add `useHTMLClass`
