+++ packages/@vuepress/theme-default/src/client/composables/index.ts
@@ -1,4 +1,5 @@
 export * from './useDarkMode'
+export * from './useHTMLClass'
 export * from './useNavLink'
 export * from './usePrefersColorScheme'
 export * from './useResolveRouteWithRedirect'
